reorganize code and delete old pipeline
